♻️ highlighttext to component
